// src/Components/Shared/Types.ts
export enum SelectedPage {
  Home = 'home',
  Specialites = 'specialites',
  PlateauTechnique = 'plateautechnique',
  Localisation = 'localisation',
  Contact = 'contact',
}

export enum SelectedService {
  ChirurgieGenerale = 'chirurgiegenerale',
  ObstetriqueGynecologie = 'obstetriquegynecologie',
  Pediatrie = 'pediatrie',
  CardiologieEtChirurgie = 'cardiologieetchirurgie',
  ChirurgieUrologique = 'chirurgieurologique',
  ChirurgieMaxilloFaciale = 'chirurgiemaxillofaciale',
  Ophtalmologie = 'ophtalmologie',
  ORL = 'orl',
  Traumatologie = 'traumatologie',
  ChirurgiePlastique = 'chirurgieplastique',
  MedecineNucleaire = 'medecinenucleaire',
  CardiologieInterventionnelle = 'cardiologieinterventionnelle',
}